+++ addon/services/oauth-service.js
@@ -4,10 +4,9 @@ import serviceMixin from '../mixins/service-mixin';
 export default Ember.Service.extend(serviceMixin, {
 
   // register an application for OAuth access
-  registerApp (itemId, redirectUris, appType = 'browser') {
-    let portalRestUrl = this.get('portalRestUrl');
-    let url = `${portalRestUrl}/oauth2/registerApp?f=json`;
-    let options = {
+  registerApp (itemId, redirectUris, appType = 'browser', portalOpts) {
+    const urlPath = `/oauth2/registerApp?f=json`;
+    const options = {
       method: 'POST',
       data: {
         itemId,
@@ -15,21 +14,20 @@ export default Ember.Service.extend(serviceMixin, {
         redirect_uris: JSON.stringify(redirectUris)
       }
     };
-    return this.request(url, options);
+    return this.request(urlPath, options, portalOpts);
   },
 
   // update an app's OAuth registration
   // TODO: have this take other params that can be updated as options
-  updateApp (clientId, redirectUris) {
-    let portalRestUrl = this.get('portalRestUrl');
-    let url = `${portalRestUrl}/oauth2/apps/${clientId}/update?f=json`;
-    let options = {
+  updateApp (clientId, redirectUris, portalOpts) {
+    const urlPath = `/oauth2/apps/${clientId}/update?f=json`;
+    const options = {
       method: 'POST',
       data: {
         client_id: clientId,
         redirect_uris: JSON.stringify(redirectUris)
       }
     };
-    return this.request(url, options);
+    return this.request(urlPath, options, portalOpts);
   }
 });