+++ addon/services/portal-service.js
@@ -4,15 +4,24 @@ import serviceMixin from '../mixins/service-mixin';
 export default Ember.Service.extend(serviceMixin, {
   session: Ember.inject.service('session'),
 
+  getById (id, portalOpts) {
+    const urlPath = `/portals/${id}?f=json`;
+    return this.request(urlPath, null, portalOpts);
+  },
+
+  self (portalOpts) {
+    const urlPath = `/portals/self?f=json`;
+    return this.request(urlPath, null, portalOpts);
+  },
+
   /**
    * Update the portal
    */
-  update (portal) {
+  update (portal, portalOpts) {
     // console.log('Portal Service got update for ' + portal.id);
-    let portalRestUrl = this.get('portalRestUrl');
-    let url = `${portalRestUrl}/portals/${portal.id}/update?f=json`;
-    let serializedPortal = this._serializePortal(portal);
-    return this._post(url, serializedPortal);
+    const urlPath = `/portals/${portal.id}/update?f=json`;
+    const serializedPortal = this._serializePortal(portal);
+    return this._post(urlPath, serializedPortal, portalOpts);
   },
 
   /**
@@ -21,36 +30,41 @@ export default Ember.Service.extend(serviceMixin, {
    * we strip it down A LOT.
    */
   _serializePortal (portal) {
-    let clone = {};
-    // if more properties are needed, please open a PR on this project
+    const allowedProperties = [ 'access', 'creditAssignments' ];
+
+    let result = allowedProperties.reduce((acc, property) => {
+      if (portal.hasOwnProperty(property)) {
+        acc[property] = portal[property];
+      }
+      return acc;
+    }, {});
+
     if (portal.portalProperties) {
-      clone.portalProperties = JSON.stringify(portal.portalProperties);
+      result.portalProperties = JSON.stringify(portal.portalProperties);
     }
 
-    return clone;
+    return result;
   },
 
   /**
    * Shared logic for POST operations
    */
-  _post (url, obj) {
-    let options = {
+  _post (urlPath, obj, portalOpts) {
+    const options = {
       method: 'POST',
       data: obj
     };
-    return this.request(url, options);
+    return this.request(urlPath, options, portalOpts);
   },
 
   /**
    * Upload a resource (file) to an item
    */
-  uploadResource (file) {
+  uploadResource (file, portalOpts) {
     // Valid types
     // const validTypes = ['json', 'xml', 'txt', 'png', 'jpeg', 'gif', 'bmp', 'pdf', 'mp3', 'mp4', 'zip'];
     // TODO: Check type
-    // const portalId = this.get('session.portal.id');
-    let portalRestUrl = this.get('portalRestUrl');
-    let url = `${portalRestUrl}/portals/self/addresource?f=json`;
+    const urlPath = `/portals/self/addresource?f=json`;
     let options = {};
     options.body = new FormData();
     // stuff the file into the formData...
@@ -58,53 +72,61 @@ export default Ember.Service.extend(serviceMixin, {
     options.body.append('text', null);
     options.body.append('key', file.name);
     options.method = 'POST';
-    return this.request(url, options);
+    return this.request(urlPath, options, portalOpts);
   },
 
   /**
    * Add a resource
    */
-  addResource (name, content) {
-    // const portalId = this.get('session.portal.id');
-    let portalRestUrl = this.get('portalRestUrl');
-    let url = `${portalRestUrl}/portals/self/addresource?f=json`;
-    let options = {
+  addResource (name, content, portalOpts) {
+    const urlPath = `/portals/self/addresource?f=json`;
+    const options = {
       method: 'POST',
       data: {
         key: name,
         text: content
       }
     };
-    return this.request(url, options);
+    return this.request(urlPath, options, portalOpts);
   },
 
   /**
    * Get the resources associated with an Item
    */
-  getResources () {
-    // const portalId = this.get('session.portal.id');
-    let portalRestUrl = this.get('portalRestUrl');
-    let url = `${portalRestUrl}/portals/self/resources?f=json`;
-    return this.request(url);
+  getResources (portalOpts) {
+    const urlPath = `/portals/self/resources?f=json`;
+    return this.request(urlPath, null, portalOpts);
   },
 
   /**
    * Remove a resource
    */
-  removeResource (resourceName) {
-    // const portalId = this.get('session.portal.id');
-    let portalRestUrl = this.get('portalRestUrl');
-    let url = `${portalRestUrl}/portals/self/removeresource?f=json`;
-    return this.request(url, {method: 'POST', data: {key: resourceName}});
+  removeResource (resourceName, portalOpts) {
+    const urlPath = `/portals/self/removeresource?f=json`;
+    return this.request(urlPath, { method: 'POST', data: { key: resourceName } }, portalOpts);
   },
 
   /**
   * Paged access to users in a portal
   */
-  users (portalId, start = 1, num = 100) {
-    const portalBaseUrl = this.get('portalRestUrl');
-    let url = `${portalBaseUrl}/portals/${portalId}/users/?f=json&start=${start}&num=${num}`;
-    return this.request(url);
+  users (portalId, start = 1, num = 100, portalOpts) {
+    const urlPath = `/portals/${portalId}/users/?f=json&start=${start}&num=${num}`;
+    return this.request(urlPath, null, portalOpts);
+  },
+
+  configureSocialProviders (opts, portalOpts) {
+    /*
+      portalOpts: {
+        signUpMode: Automatic | ???
+        providers: facebook,google
+        level: 1 | 2
+        role (optional): org_user (default) | org_publisher | id of custom role
+        userCreditAssignment (optional): <number>
+        groups (optional): groupId1, groupId2, ...
+      }
+    */
+    const urlPath = `/portals/self/socialProviders/configure?f=json`;
+    return this._post(urlPath, opts, portalOpts);
   }
 
 });