+++ tests/dummy/app/users/index/controller.js
@@ -2,11 +2,11 @@ import Ember from 'ember';
 
 export default Ember.Controller.extend({
 
-  itemsService: Ember.inject.service('items-service'),
+  userService: Ember.inject.service(),
 
   session: Ember.inject.service(),
 
-  queryParams: [ 'start', 'num', 'q' ],
+  queryParams: [ 'start', 'num', 'q', 'token', 'portalHostname' ],
 
   start: 1,
 
@@ -15,6 +15,14 @@ export default Ember.Controller.extend({
   query: '*',
 
   num: 10,
+
+  portalHostname: '',
+
+  portalHostnameCopy: Ember.computed.reads('portalHostname'),
+
+  token: '',
+
+  tokenCopy: Ember.computed.reads('token'),
 
   totalCount: Ember.computed('model.total', function () {
     return this.get('model.total');
@@ -35,8 +43,14 @@ export default Ember.Controller.extend({
       this.set('q', this.get('query'));
       // reset the page
       this.set('start', 1);
-      this.transitionToRoute('items.index');
+      const queryParams = {
+        q: this.get('query'),
+        start: 1,
+        num: this.get('num'),
+        portalHostname: this.get('portalHostnameCopy'),
+        token: this.get('tokenCopy')
+      };
+      this.transitionToRoute('users', { queryParams: queryParams });
     }
-
   }
 });